Etapa 2: que Atenea pueda existir fuera de una maquina
El backend estaba bien construido por dentro y completamente desconectado del
mundo por fuera. Esto cierra esa brecha.

Arranque que se niega a ir mal configurado. En produccion, un validador exige un
secreto de firma propio y largo, un proveedor de IA real con su clave, embeddings
reales y una lista de origenes CORS. Si falta algo, la aplicacion no levanta y
dice cual. Antes arrancaba con el secreto de desarrollo, que esta publicado en
este repositorio, y con el proveedor simulado: habria servido lecciones
inventadas firmando tokens que cualquiera puede falsificar.

Salud en las dos rutas. `/api/v1/health` es la del contrato; `/health` es la que
esperan las plataformas, que no saben del prefijo de version. El healthcheck del
despliegue apuntaba a la segunda, que devolvia 404: habria puesto el servicio en
un bucle de reinicios con la aplicacion sana. Ademas responde 503 cuando la base
no contesta, en vez de decir que todo va bien.

Freno de peticiones. `slowapi` llevaba tiempo en las dependencias, tres routers
tenian un TODO esperandolo y `RATE_LIMITED` era el unico codigo del catalogo que
nadie lanzaba. Ahora estan los cuatro limites del §8.7 mas uno en la subida, y el
acceso tiene el suyo: probar contrasenas en masa costaba un bcrypt de coste 12
por intento sobre un backend sincrono, que es un robo y una denegacion de
servicio a la vez. Se aplican como dependencia y no como decorador, porque el
envoltorio del decorador deja a FastAPI sin resolver los tipos de las rutas.

Reportar contenido ya no es un arma. No comprobaba propiedad y un solo reporte
retiraba la pregunta para todo el mundo: una cuenta recien creada podia dejar la
Ruta del Reino sin preguntas. Ahora hace falta acceso al contenido, y en el
material compartido, varios reportes de personas distintas. La clave de
idempotencia ya no lleva la hora, asi que insistir cuenta una vez.

Despliegue en el formato vigente. `.railway/railway.ts` sustituye a
`railway.toml`, que Railway declaro obsoleto y no admite en servicios nuevos, y
cuya mitad (worker, variables, volumen) estaba escrita solo como comentario. Las
migraciones y las semillas van antes del arranque: sin sembrar, una base recien
migrada queda sin configuracion, sin niveles, sin objetos y sin misiones.

Un servicio, a proposito: el material del aprendiz vive en disco y un volumen se
monta en un solo servicio, asi que una API y un worker separados no compartirian
los archivos. El procesador de trabajos corre dentro de la API
(`WORKER_EN_PROCESO`). Cuando el material viva en un bucket, se apaga y el worker
vuelve a ser un servicio aparte.

Integracion continua, que no existia: migraciones de cero a head, deriva entre
modelos y base, estilo y las dos suites. Para que `ruff check` pueda ser un
guardian se dejo el backend limpio, incluida la regla que da sentido a las
marcas de importacion perezosa del §1.3.

De paso, la suite dejo de romperse al final: varias suites abren una conexion
para toda la ejecucion y el tope de transaccion ociosa se quedo corto al crecer
las pruebas. El fallo aparecia como un corte de red en el ultimo caso, que no
tenia nada que ver.

498 pruebas de backend y 29 de cliente en verde. Comprobado contra la API
levantada: el freno responde 429 con el sobre del contrato a la undecima entrada,
y una cuenta que no ha tocado la Ruta recibe 404 al intentar reportarla, sin
dejar rastro en la pregunta.

+++ backend/app/modules/ingestion/fragmentacion.py
@@ -105,7 +105,7 @@ class ParametrosTroceo:
     @property
     def tokens_solape(self) -> int:
         """Tokens que se arrastran de un fragmento de prosa al siguiente."""
-        return max(0, int(round(self.target_tokens * self.overlap_pct)))
+        return max(0, round(self.target_tokens * self.overlap_pct))
 
 
 def parametros_de(configuracion: dict | None) -> ParametrosTroceo:
@@ -285,7 +285,7 @@ def analizar_bloques(texto: str) -> list[Bloque]:
 # ---------------------------------------------------------------------------
 
 
-def _unidades_de_corte(texto: str, maximo: int) -> list[str]:
+def _unidades_de_corte(texto: str, maximo: int) -> list[str]:  # noqa: ARG001 - firma fijada por quien llama
     """Unidades por las que se puede partir un párrafo: frases y, si no hay, palabras.
 
     Un PDF mal extraído produce párrafos kilométricos sin un solo punto. Cortarlos por
